# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NCKH SaaS Multi-tenancy API
  version: 0.1.0
  description: >-
    Hợp đồng REST cho control plane và application plane. Các API nghiệp vụ lấy
    tenant từ host và JWT; không nhận tenant_id từ payload để phân quyền.
servers:
  - url: http://accounts.localhost:8080/api/v1
    description: Control host local
  - url: http://{tenant}.localhost:8080/api/v1
    description: Tenant host local
    variables:
      tenant:
        default: pool-demo
security:
  - bearerAuth: []
...
paths:
...
  /dashboard:
    get:
      tags: [Projects]
      operationId: dashboard
      responses:
        '200':
          description: Metrics restricted to projects visible to current user
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DashboardView'}
        default: {$ref: '#/components/responses/Error'}
  /projects:
    get:
      tags: [Projects]
      operationId: listProjects
      responses:
        '200':
          description: Projects with current user's project membership
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ProjectView'}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Projects]
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateProjectRequest'}
      responses:
        '201':
          description: Project, manager membership and default Kanban board created atomically
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectView'}
        default: {$ref: '#/components/responses/Error'}
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    put:
      tags: [Projects]
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateProjectRequest'}
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectView'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      tags: [Projects]
      operationId: deleteProject
      responses:
        '204': {description: Project soft-deleted with its content retained for audit}
        default: {$ref: '#/components/responses/Error'}
  /projects/{projectId}/status:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    patch:
      tags: [Projects]
      operationId: changeProjectStatus
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChangeProjectStatusRequest'}
      responses:
        '200':
          description: Project archived or restored by a project manager
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectView'}
        default: {$ref: '#/components/responses/Error'}
  /projects/{projectId}/members:
    get:
      tags: [Projects]
      operationId: listProjectMembers
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project memberships
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ProjectMemberView'}
        default: {$ref: '#/components/responses/Error'}
  /projects/{projectId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/UserId'
    put:
      tags: [Projects]
      operationId: setProjectMember
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SetProjectMemberRequest'}
      responses:
        '200':
          description: Project membership created or updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectMemberView'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      tags: [Projects]
      operationId: removeProjectMember
      responses:
        '204': {description: Project membership removed}
        default: {$ref: '#/components/responses/Error'}
  /projects/{projectId}/boards:
    get:
      tags: [Projects]
      operationId: listProjectBoards
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Active boards visible through project membership
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/BoardSummaryView'}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Projects]
      operationId: createBoard
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateBoardRequest'}
      responses:
        '201':
          description: Board with default columns
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BoardView'}
        default: {$ref: '#/components/responses/Error'}
  /boards/{boardId}:
    parameters:
      - $ref: '#/components/parameters/BoardId'
    get:
      tags: [Projects]
      operationId: getBoard
      responses:
        '200':
          description: Board, columns and tasks
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BoardView'}
        default: {$ref: '#/components/responses/Error'}
    put:
      tags: [Projects]
      operationId: updateBoard
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateBoardRequest'}
      responses:
        '200':
          description: Board renamed with optimistic version check
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BoardView'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      tags: [Projects]
      operationId: deleteBoard
      responses:
        '204': {description: Board soft-deleted}
        default: {$ref: '#/components/responses/Error'}
  /boards/{boardId}/columns:
    post:
      tags: [Projects]
      operationId: createColumn
      parameters:
        - $ref: '#/components/parameters/BoardId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateColumnRequest'}
      responses:
        '201':
          description: Column created and board version advanced
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BoardView'}
        default: {$ref: '#/components/responses/Error'}
  /boards/{boardId}/columns/order:
    put:
      tags: [Projects]
      operationId: reorderColumns
      parameters:
        - $ref: '#/components/parameters/BoardId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReorderColumnsRequest'}
      responses:
        '200':
          description: Columns reordered atomically; 409 is returned for a stale board version
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BoardView'}
        default: {$ref: '#/components/responses/Error'}
  /boards/{boardId}/columns/{columnId}:
    parameters:
      - $ref: '#/components/parameters/BoardId'
      - $ref: '#/components/parameters/ColumnId'
    patch:
      tags: [Projects]
      operationId: updateColumn
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateColumnRequest'}
      responses:
        '200':
          description: Column renamed and board version advanced
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BoardView'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      tags: [Projects]
      operationId: deleteColumn
      parameters:
        - name: version
          in: query
          required: true
          schema: {type: integer, format: int64, minimum: 0}
      responses:
        '200':
          description: Empty, non-final column deleted and board version advanced
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BoardView'}
        default: {$ref: '#/components/responses/Error'}
  /boards/{boardId}/tasks:
    post:
      tags: [Projects]
      operationId: createTask
      parameters:
        - $ref: '#/components/parameters/BoardId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateTaskRequest'}
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskView'}
        default: {$ref: '#/components/responses/Error'}
  /boards/{boardId}/tasks/{taskId}/position:
    patch:
      tags: [Projects]
      operationId: moveTask
      parameters:
        - $ref: '#/components/parameters/BoardId'
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/MoveTaskRequest'}
      responses:
        '200':
          description: Task moved; 409 is returned for stale version
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskView'}
        default: {$ref: '#/components/responses/Error'}
  /boards/{boardId}/tasks/order:
    put:
      tags: [Projects]
      operationId: reorderTasks
      parameters:
        - $ref: '#/components/parameters/BoardId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReorderTasksRequest'}
      responses:
        '200':
          description: Task positions updated atomically; any stale version rolls back the batch
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BoardView'}
        default: {$ref: '#/components/responses/Error'}
  /tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      tags: [Projects]
      operationId: getTask
      responses:
        '200':
          description: Active task visible through project membership
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskView'}
        default: {$ref: '#/components/responses/Error'}
    patch:
      tags: [Projects]
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateTaskRequest'}
      responses:
        '200':
          description: Task updated; 409 is returned for stale version
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskView'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      tags: [Projects]
      operationId: deleteTask
      responses:
        '204': {description: Task and its one-level subtasks soft-deleted}
        default: {$ref: '#/components/responses/Error'}
  /tasks/{taskId}/comments:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      tags: [Projects]
      operationId: listComments
      responses:
        '200':
          description: Task comments
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/CommentView'}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Projects]
      operationId: createComment
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateCommentRequest'}
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommentView'}
        default: {$ref: '#/components/responses/Error'}
  /comments/{commentId}:
    parameters:
      - name: commentId
        in: path
        required: true
        schema: {$ref: '#/components/schemas/UUID'}
    patch:
      tags: [Projects]
      operationId: updateComment
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateCommentRequest'}
      responses:
        '200':
          description: Comment updated by its author or moderated by a project manager
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommentView'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      tags: [Projects]
      operationId: deleteComment
      responses:
        '204': {description: Comment soft-deleted by its author or a project manager}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserId:
      name: userId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
    BoardId:
      name: boardId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
    ColumnId:
      name: columnId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
    TaskId:
      name: taskId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
...
  responses:
    Error:
      description: Structured API error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
  schemas:
    UUID: {type: string, format: uuid}
    Instant: {type: string, format: date-time}
...
    ProjectRole: {type: string, enum: [MANAGER, MEMBER, VIEWER]}
    ProjectStatus: {type: string, enum: [ACTIVE, ARCHIVED, DELETED]}
...
    ApiError:
      type: object
      required: [timestamp, status, code, message, path, fieldErrors]
      properties:
        timestamp: {$ref: '#/components/schemas/Instant'}
        status: {type: integer}
        code: {type: string}
        message: {type: string}
        path: {type: string}
        fieldErrors: {type: object, additionalProperties: {type: string}}
        requestId: {type: [string, 'null']}
...
    DashboardView:
      type: object
      required: [projects, boards, openTasks, overdueTasks, unreadNotifications]
      properties:
        projects: {type: integer, format: int64}
        boards: {type: integer, format: int64}
        openTasks: {type: integer, format: int64}
        overdueTasks: {type: integer, format: int64}
        unreadNotifications: {type: integer, format: int64}
    ProjectView:
      type: object
      required: [id, name, status, createdBy, createdAt, updatedAt, role, memberCount, taskCount, completedTaskCount]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        name: {type: string}
        description: {type: [string, 'null']}
        status: {$ref: '#/components/schemas/ProjectStatus'}
        createdBy: {$ref: '#/components/schemas/UUID'}
        createdAt: {$ref: '#/components/schemas/Instant'}
        updatedAt: {$ref: '#/components/schemas/Instant'}
        role: {$ref: '#/components/schemas/ProjectRole'}
        boardId: {anyOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        memberCount: {type: integer, format: int64}
        taskCount: {type: integer, format: int64}
        completedTaskCount: {type: integer, format: int64}
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name: {type: string, maxLength: 160}
        description: {type: [string, 'null'], maxLength: 5000}
    UpdateProjectRequest:
      $ref: '#/components/schemas/CreateProjectRequest'
    ChangeProjectStatusRequest:
      type: object
      required: [status]
      properties:
        status: {$ref: '#/components/schemas/ProjectStatus'}
    ProjectMemberView:
      type: object
      required: [projectId, userId, role]
      properties:
        projectId: {$ref: '#/components/schemas/UUID'}
        userId: {$ref: '#/components/schemas/UUID'}
        role: {$ref: '#/components/schemas/ProjectRole'}
    SetProjectMemberRequest:
      type: object
      required: [role]
      properties: {role: {$ref: '#/components/schemas/ProjectRole'}}
    ColumnView:
      type: object
      required: [id, name, position]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        name: {type: string}
        position: {type: number}
    TaskView:
      type: object
      required: [id, projectId, boardId, columnId, title, position, version, createdAt, updatedAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        projectId: {$ref: '#/components/schemas/UUID'}
        boardId: {$ref: '#/components/schemas/UUID'}
        columnId: {$ref: '#/components/schemas/UUID'}
        parentTaskId: {anyOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        title: {type: string}
        description: {type: [string, 'null']}
        assigneeUserId: {anyOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        dueAt: {anyOf: [{$ref: '#/components/schemas/Instant'}, {type: 'null'}]}
        position: {type: number}
        version: {type: integer, format: int64}
        createdAt: {$ref: '#/components/schemas/Instant'}
        updatedAt: {$ref: '#/components/schemas/Instant'}
    BoardView:
      type: object
      required: [id, projectId, name, version, columns, tasks]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        projectId: {$ref: '#/components/schemas/UUID'}
        name: {type: string}
        version: {type: integer, format: int64, minimum: 0}
        columns: {type: array, items: {$ref: '#/components/schemas/ColumnView'}}
        tasks: {type: array, items: {$ref: '#/components/schemas/TaskView'}}
    BoardSummaryView:
      type: object
      required: [id, projectId, name, version, createdAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        projectId: {$ref: '#/components/schemas/UUID'}
        name: {type: string}
        version: {type: integer, format: int64, minimum: 0}
        createdAt: {$ref: '#/components/schemas/Instant'}
    CreateBoardRequest:
      type: object
      required: [name]
      properties: {name: {type: string, maxLength: 160}}
    UpdateBoardRequest:
      type: object
      required: [name, version]
      properties:
        name: {type: string, maxLength: 160}
        version: {type: integer, format: int64, minimum: 0}
    CreateColumnRequest:
      type: object
      required: [name, version]
      properties:
        name: {type: string, maxLength: 120}
        version: {type: integer, format: int64, minimum: 0}
    UpdateColumnRequest:
      $ref: '#/components/schemas/CreateColumnRequest'
    ReorderColumnsRequest:
      type: object
      required: [columnIds, version]
      properties:
        columnIds:
          type: array
          minItems: 1
          uniqueItems: true
          items: {$ref: '#/components/schemas/UUID'}
        version: {type: integer, format: int64, minimum: 0}
    CreateTaskRequest:
      type: object
      required: [columnId, title]
      properties:
        columnId: {$ref: '#/components/schemas/UUID'}
        parentTaskId: {anyOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        title: {type: string, maxLength: 240}
        description: {type: [string, 'null'], maxLength: 10000}
        assigneeUserId: {anyOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        dueAt: {anyOf: [{$ref: '#/components/schemas/Instant'}, {type: 'null'}]}
        position: {type: [number, 'null']}
    UpdateTaskRequest:
      type: object
      required: [columnId, title, version]
      properties:
        columnId: {$ref: '#/components/schemas/UUID'}
        title: {type: string, maxLength: 240}
        description: {type: [string, 'null'], maxLength: 10000}
        assigneeUserId: {anyOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        dueAt: {anyOf: [{$ref: '#/components/schemas/Instant'}, {type: 'null'}]}
        position: {type: [number, 'null']}
        version: {type: integer, format: int64}
    MoveTaskRequest:
      type: object
      required: [targetColumnId, targetPosition, version]
      properties:
        targetColumnId: {$ref: '#/components/schemas/UUID'}
        targetPosition: {type: number}
        version: {type: integer, format: int64}
    ReorderTaskItem:
      type: object
      required: [taskId, targetColumnId, targetPosition, version]
      properties:
        taskId: {$ref: '#/components/schemas/UUID'}
        targetColumnId: {$ref: '#/components/schemas/UUID'}
        targetPosition: {type: number}
        version: {type: integer, format: int64, minimum: 0}
    ReorderTasksRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items: {$ref: '#/components/schemas/ReorderTaskItem'}
    CommentView:
      type: object
      required: [id, taskId, authorUserId, body, createdAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        taskId: {$ref: '#/components/schemas/UUID'}
        authorUserId: {$ref: '#/components/schemas/UUID'}
        body: {type: string}
        createdAt: {$ref: '#/components/schemas/Instant'}
    CreateCommentRequest:
      type: object
      required: [body]
      properties: {body: {type: string, maxLength: 10000}}
    UpdateCommentRequest:
      $ref: '#/components/schemas/CreateCommentRequest'
...